requirements.txt
python-dotenv
azure-storage-blob
azure-storage-file-share
azure-identity
requests
openpyxl